Deprecate the use of SUBPROCESS_NUM in Makefile
GNU Make can handle the maximum number of subprocesses correctly by
using the objserver feature, it is enabled by default if we call another
make process using `$(MAKE)`. So we don't need to set the SUBPROCESS_NUM
variable in Makefile. Using `-j` on the top level `make` to control the
number of subprocesses more flexible and safer.

Reference:
[1] https://www.gnu.org/software/make/manual/html_node/POSIX-Jobserver.html
[2] https://www.gnu.org/software/make/manual/html_node/MAKE-Variable.html

Signed-off-by: Yangyu Chen <cyy@cyyself.name>