Makefile: Add TAG option to specify the suffix name of the output file
Now, we can use the $TAG option to specify the suffix name of the output
file. This is useful when we want to run the same benchmark with
different compiler options such as different -march, -mtune or PGO.

For example, we can run the following command to build the default
target and the target with pgo:

make build-all -j `nproc`
make TAG="pgo" build-allpgo -j `nproc`

Then, each benchmark will have two executables: the default one and the
one with the suffix "pgo". The suffix name is specified by the $TAG.

./400.perlbench/build
├── 400.perlbench
├── 400.perlbenchpgo
├── run-train.log
├── pgosrc
└── src

This is useful when we want to compare the performance of different
compiler options.